requirements.txt
Flask==0.12.2
numpy==1.13.3
Pillow==5.0.0
Werkzeug==0.14.1

